fix static issues
