+++ dev/tests/api-functional/testsuite/Magento/Catalog/Api/ProductRepositoryInterfaceTest.php
@@ -48,10 +48,8 @@ class ProductRepositoryInterfaceTest extends WebapiAbstract
     private const SERVICE_NAME = 'catalogProductRepositoryV1';
     private const SERVICE_VERSION = 'V1';
     private const RESOURCE_PATH = '/V1/products';
-
     private const KEY_TIER_PRICES = 'tier_prices';
     private const KEY_SPECIAL_PRICE = 'special_price';
-    private const KEY_CATEGORY_LINKS = 'category_links';
 
     /**
      * @var array
@@ -1659,86 +1657,6 @@ class ProductRepositoryInterfaceTest extends WebapiAbstract
             StockItemInterface::IS_DECIMAL_DIVIDED => 0,
             StockItemInterface::STOCK_STATUS_CHANGED_AUTO => 0,
         ];
-    }
-
-    /**
-     * Test product category links
-     *
-     * @magentoApiDataFixture Magento/Catalog/_files/category_product.php
-     */
-    public function testProductCategoryLinks()
-    {
-        // Create simple product
-        $productData = $this->getSimpleProductData();
-        $productData[ProductInterface::EXTENSION_ATTRIBUTES_KEY] = [
-            self::KEY_CATEGORY_LINKS => [['category_id' => 333, 'position' => 0]],
-        ];
-        $response = $this->saveProduct($productData);
-        $this->assertEquals(
-            [['category_id' => 333, 'position' => 0]],
-            $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY][self::KEY_CATEGORY_LINKS]
-        );
-        $response = $this->getProduct($productData[ProductInterface::SKU]);
-        $this->assertArrayHasKey(ProductInterface::EXTENSION_ATTRIBUTES_KEY, $response);
-        $extensionAttributes = $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY];
-        $this->assertArrayHasKey(self::KEY_CATEGORY_LINKS, $extensionAttributes);
-        $this->assertEquals([['category_id' => 333, 'position' => 0]], $extensionAttributes[self::KEY_CATEGORY_LINKS]);
-    }
-
-    /**
-     * Test update product category without categories
-     *
-     * @magentoApiDataFixture Magento/Catalog/_files/category_product.php
-     */
-    public function testUpdateProductCategoryLinksNullOrNotExists()
-    {
-        $response = $this->getProduct('simple333');
-        // update product without category_link or category_link is null
-        $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY][self::KEY_CATEGORY_LINKS] = null;
-        $response = $this->updateProduct($response);
-        $this->assertEquals(
-            [['category_id' => 333, 'position' => 0]],
-            $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY][self::KEY_CATEGORY_LINKS]
-        );
-        unset($response[ProductInterface::EXTENSION_ATTRIBUTES_KEY][self::KEY_CATEGORY_LINKS]);
-        $response = $this->updateProduct($response);
-        $this->assertEquals(
-            [['category_id' => 333, 'position' => 0]],
-            $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY][self::KEY_CATEGORY_LINKS]
-        );
-    }
-
-    /**
-     * Test update product category links position
-     *
-     * @magentoApiDataFixture Magento/Catalog/_files/category_product.php
-     */
-    public function testUpdateProductCategoryLinksPosistion()
-    {
-        $response = $this->getProduct('simple333');
-        // update category_link position
-        $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY][self::KEY_CATEGORY_LINKS] = [
-            ['category_id' => 333, 'position' => 10],
-        ];
-        $response = $this->updateProduct($response);
-        $this->assertEquals(
-            [['category_id' => 333, 'position' => 10]],
-            $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY][self::KEY_CATEGORY_LINKS]
-        );
-    }
-
-    /**
-     * Test update product category links unassing
-     *
-     * @magentoApiDataFixture Magento/Catalog/_files/category_product.php
-     */
-    public function testUpdateProductCategoryLinksUnassign()
-    {
-        $response = $this->getProduct('simple333');
-        // unassign category_links from product
-        $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY][self::KEY_CATEGORY_LINKS] = [];
-        $response = $this->updateProduct($response);
-        $this->assertArrayNotHasKey(self::KEY_CATEGORY_LINKS, $response[ProductInterface::EXTENSION_ATTRIBUTES_KEY]);
     }
 
     /**
